adds permission to view logger url
Adds permission `spark.viewurl` that can be given or revoked. This is useful because it will allow regular non-opped players to run the `/spark profiler` command and generate a profiler report, but it will not show them the results URL if the permission isn't given, allowing users to create detailed bug reports without an admin needing to be active.