Optimise WorldStatisticsProvider regionising
By using BFS and a chunk coordinate->chunk map, we can regionise
in linear time - as opposed to quadratic. We also can remove the
merging logic, as using a BFS guarantees that all adjacent chunks
are in the same region.

I noticed that this part of the code was taking excessive time
while testing Moonrise on 1.21.5 with ~100k chunks loaded (using
the larger view distance startup flag), as the old regionising code
was quadratic (>60% of the upload time). With the new regionising
code, it was reduced to <1% for the upload time.

The adjacent distance checks are identical to the old code, so this
will produce the exact same regions as before.

If fastutil were in the common package, we could remove the added
Coordinate class and just the long key directly. Using a Long in
HashMap is not a solution, as the hashcode becomes x ^ z which has
many duplicates.